LPS-55528 Move store tests to their modules
