+++ modules/portal/portal-store-cmis/ivy.xml
@@ -17,6 +17,7 @@
 	</publications>
 
 	<dependencies defaultconf="default">
+		<dependency conf="test->default" name="com.liferay.arquillian.extension.junit.bridge" org="com.liferay" rev="1.0.0-SNAPSHOT" />
 		<dependency name="chemistry-opencmis-client-api" org="org.apache.chemistry.opencmis" rev="0.13.0"/>
 		<dependency name="chemistry-opencmis-client-bindings" org="org.apache.chemistry.opencmis" rev="0.13.0"/>
 		<dependency name="chemistry-opencmis-client-impl" org="org.apache.chemistry.opencmis" rev="0.13.0"/>
